+++ examples/basic-local/examples/view-examples.rs
@@ -3,11 +3,9 @@ use bonsaidb::{
         connection::Connection,
         document::CollectionDocument,
         schema::{
-            view::CollectionViewSchema, Collection, CollectionName, DefaultSerialization,
-            DefaultViewSerialization, Name, ReduceResult, Schematic, SerializedCollection, View,
-            ViewMapResult, ViewMappedValue,
+            view::CollectionViewSchema, Collection, DefaultViewSerialization, Name, ReduceResult,
+            SerializedCollection, View, ViewMapResult, ViewMappedValue,
         },
-        Error,
     },
     local::{
         config::{Builder, StorageConfiguration},
@@ -17,22 +15,11 @@ use bonsaidb::{
 use serde::{Deserialize, Serialize};
 
 // begin rustme snippet: snippet-a
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, Collection)]
+#[collection(name = "shapes", views = [ShapesByNumberOfSides])]
 struct Shape {
     pub sides: u32,
 }
-
-impl Collection for Shape {
-    fn collection_name() -> CollectionName {
-        CollectionName::private("shapes")
-    }
-
-    fn define_views(schema: &mut Schematic) -> Result<(), Error> {
-        schema.define_view(ShapesByNumberOfSides)
-    }
-}
-
-impl DefaultSerialization for Shape {}
 
 #[derive(Debug, Clone)]
 struct ShapesByNumberOfSides;